Week 5/requirements.txt
langgraph
langchain
langchain-openai
python-dotenv
ddgs
PyPDF2
pydantic
typing-extensions
streamlit
